Fixed typo while fetching "foundGreigeOrder"

+++ src/controllers/greigeController/deleteGreigeOrder.controller.js
@@ -7,7 +7,7 @@ const deleteGreigeOrder = asyncHandler(async (request, response) =>{
         throw new apiError(400, "Greige order ID is required")
     }
 
-    const foundGreigeOrder = await Greige(orderId);
+    const foundGreigeOrder = await Greige.findById(orderId);
 
     if(!foundGreigeOrder){
         throw new apiError(404, "Greige order not found maybe already deleted")
@@ -15,7 +15,10 @@ const deleteGreigeOrder = asyncHandler(async (request, response) =>{
 
     const unit3OrderId = foundGreigeOrder.unit3OrderId;
 
-    await Unit3.findByIdAndDelete(unit3OrderId);
+    if(unit3OrderId){
+        await Unit3.findByIdAndDelete(unit3OrderId);
+    }
+
     await Greige.findByIdAndDelete(orderId);
 
     return response.status(200)
